Changing buttons to a to fix ' in titles
Fixed a bug - cannot add or remove movies from watchlist that have a ' in the title

+++ server.py
@@ -49,8 +49,9 @@ def mymovies():
     if removeWishlist != "":
         c.execute("SELECT id FROM movie_database_c WHERE title = (?)", (removeWishlist,))
         movie_ID = c.fetchone()
-        c.execute("DELETE FROM Wishlist WHERE Movie_ID = (?)", (movie_ID[0],))
-        connection.commit()
+        if movie_ID != None:
+            c.execute("DELETE FROM Wishlist WHERE Movie_ID = (?)", (movie_ID[0],))
+            connection.commit()
 
     c.execute("SELECT DISTINCT m.title, m.genres, m.vote_average, m.vote_count, m.release_date, m.adult FROM movie_database_c AS m, Wishlist AS w, users AS u JOIN Wishlist movie_database_c ON w.Movie_ID = m.id JOIN Wishlist users ON w.USER_ID = u.user_id WHERE u.username = (?)", (username_cookie,))
     result = c.fetchall()
